// src/common/enums/setup-status.enum.ts
export enum SetupStatus {
  NOT_STARTED = 'NOT_STARTED',
  PAYMENT_PENDING = 'PAYMENT_PENDING',
  PAYMENT_COMPLETED = 'PAYMENT_COMPLETED',
  MEETING_PENDING = 'MEETING_PENDING',
  MEETING_SCHEDULED = 'MEETING_SCHEDULED',
  REQUIREMENT_COLLECTED = 'REQUIREMENT_COLLECTED',
  SETUP_IN_PROGRESS = 'SETUP_IN_PROGRESS',
  TESTING = 'TESTING',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED',
}